+++ org.mixedrealitytoolkit.input/Tests/Runtime/FuzzyGazeInteractorTests.cs
@@ -31,7 +31,7 @@ namespace MixedReality.Toolkit.Input.Tests
         public IEnumerator BasicFuzzyGazeTest()
         {
             // Confirm a FuzzyGazeInteractor is active in the scene
-            FuzzyGazeInteractor fuzzyGazeInteractor = FindObjectUtility.FindAnyObjectByType<FuzzyGazeInteractor>();
+            FuzzyGazeInteractor fuzzyGazeInteractor = Object.FindAnyObjectByType<FuzzyGazeInteractor>();
             Assert.IsNotNull(fuzzyGazeInteractor, "There is no active FuzzyGazeInteractor found in the scene.");
 
             // Instantiate two foreground cubes and one background cube for testing
@@ -107,7 +107,7 @@ namespace MixedReality.Toolkit.Input.Tests
         public IEnumerator FuzzyGazePrecisionTest()
         {
             // Confirm a FuzzyGazeInteractor is active in the scene and configure it for the test
-            FuzzyGazeInteractor fuzzyGazeInteractor = FindObjectUtility.FindFirstObjectByType<FuzzyGazeInteractor>();
+            FuzzyGazeInteractor fuzzyGazeInteractor = Object.FindFirstObjectByType<FuzzyGazeInteractor>();
             Assert.IsNotNull(fuzzyGazeInteractor, "There is no active FuzzyGazeInteractor found in the scene.");
             fuzzyGazeInteractor.precision = 0;
             fuzzyGazeInteractor.performAdditionalRaycast = false;
@@ -165,10 +165,10 @@ namespace MixedReality.Toolkit.Input.Tests
         public IEnumerator GazeFallbackWhenEyeGazeLosesTrackingTest()
         {
             // Confirm a FuzzyGazeInteractor is active in the scene
-            FuzzyGazeInteractor fuzzyGazeInteractor = FindObjectUtility.FindFirstObjectByType<FuzzyGazeInteractor>();
+            FuzzyGazeInteractor fuzzyGazeInteractor = Object.FindFirstObjectByType<FuzzyGazeInteractor>();
             Assert.IsNotNull(fuzzyGazeInteractor, "There is no active FuzzyGazeInteractor found in the scene.");
 
-            // Instantiate two foregound cubes and one background cube for testing
+            // Instantiate two foreground cubes and one background cube for testing
             GameObject cube1 = GameObject.CreatePrimitive(PrimitiveType.Cube);
             cube1.GetComponent<MeshRenderer>().material.color = Color.red;
             cube1.AddComponent<StatefulInteractable>();
@@ -234,10 +234,10 @@ namespace MixedReality.Toolkit.Input.Tests
         public IEnumerator GazeFallbackWhenEyeGazeIsDisabledTest()
         {
             // Confirm a FuzzyGazeInteractor is active in the scene
-            FuzzyGazeInteractor fuzzyGazeInteractor = FindObjectUtility.FindFirstObjectByType<FuzzyGazeInteractor>();
+            FuzzyGazeInteractor fuzzyGazeInteractor = Object.FindFirstObjectByType<FuzzyGazeInteractor>();
             Assert.IsNotNull(fuzzyGazeInteractor, "There is no active FuzzyGazeInteractor found in the scene.");
 
-            // Instantiate two foregound cubes and one background cube for testing
+            // Instantiate two foreground cubes and one background cube for testing
             GameObject cube1 = GameObject.CreatePrimitive(PrimitiveType.Cube);
             cube1.GetComponent<MeshRenderer>().material.color = Color.red;
             cube1.AddComponent<StatefulInteractable>();