+++ org.mixedrealitytoolkit.uxcore/Experimental/NonNativeKeyboard/NonNativeKeyboardTouchAdapter.cs
@@ -1,0 +1,39 @@
+﻿// Copyright (c) Mixed Reality Toolkit Contributors
+// Licensed under the BSD 3-Clause
+
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace MixedReality.Toolkit.UX.Experimental
+{
+    /// <summary>
+    /// Adds touch interaction to every button on the non-native keyboard.
+    /// </summary>
+    [RequireComponent(typeof(AudioSource))]
+    public class NonNativeKeyboardTouchAdapter : MonoBehaviour
+    {
+        /// <summary>
+        /// See <see cref="MonoBehaviour"/>.
+        /// </summary>
+        protected void Awake()
+        {
+            GetComponent<AudioSource>().playOnAwake = false;
+        }
+
+        /// <summary>
+        /// See <see cref="MonoBehaviour"/>.
+        /// </summary>
+        protected void Start()
+        {
+            var buttons = GetComponentsInChildren<Button>(true);
+            foreach (var button in buttons)
+            {
+                // The search box has an incorrect collider and should not act as a button anyway
+                if (button.gameObject.name != "search")
+                {
+                    button.gameObject.EnsureComponent<NonNativeKeyTouchAdapter>();
+                }
+            }
+        }
+    }
+}