+++ org.mixedrealitytoolkit.uxcore/SeeItSayIt/SeeItSayItLabelEnabler.cs
@@ -6,6 +6,9 @@ using UnityEngine;
 #if MRTK_INPUT_PRESENT && MRTK_SPEECH_PRESENT
 using MixedReality.Toolkit.Input;
 #endif
+#if UNITY_LOCALIZATION_PRESENT
+using UnityEngine.Localization;
+#endif
 
 namespace MixedReality.Toolkit.UX
 {
@@ -16,15 +19,25 @@ namespace MixedReality.Toolkit.UX
     /// <remarks>
     /// A "see-it say-it" label is used for accessibility and voice input. The label displays the keyword
     /// that can be spoken to active or click the associated <see cref="PressableButton"/>.
-    /// 
+    ///
     /// This class will only enable a "see-it say-it" label if the application has included the MRTK input
-    /// package, and has an active <see cref="SpeechInteractor"/> object when this component's 
+    /// package, and has an active <see cref="SpeechInteractor"/> object when this component's
     /// <see cref="Start"/> method is invoked.
     /// </remarks>
     [RequireComponent(typeof(PressableButton))]
     [AddComponentMenu("MRTK/UX/See It Say It Label")]
     public class SeeItSayItLabelEnabler : MonoBehaviour
     {
+        /// <summary>
+        /// The <see cref="PressableButton"/> present on the same GameObject.
+        /// </summary>
+        private PressableButton pressableButton;
+
+        /// <summary>
+        /// The <see cref="TMP_Text"/> used to display the label present in child.
+        /// </summary>
+        private TMP_Text labelText;
+
         [SerializeField]
         [Tooltip("The GameObject for the see-it say-it label to be enabled.")]
         private GameObject seeItSayItLabel;
@@ -36,6 +49,32 @@ namespace MixedReality.Toolkit.UX
         {
             get => seeItSayItLabel;
             set => seeItSayItLabel = value;
+        }
+
+#if UNITY_LOCALIZATION_PRESENT
+        [SerializeField]
+        [Tooltip("The LocalizedString that define the label pattern. Use a smart string with one argument that will be replaced by the button's speech recognition keyword (e.g: \"Say '{0}'\").")]
+        private LocalizedString localizedPattern;
+#endif
+
+        [SerializeField]
+        [Tooltip("The pattern for the see-it say-it label using string.Format() when localization isn't used.")]
+        private string pattern = "Say '{0}'";
+
+        /// <summary>
+        /// The pattern for the see-it say-it label using string.Format() when localization isn't used.
+        /// </summary>
+        public string Pattern
+        {
+            get => pattern;
+            set
+            {
+                pattern = value;
+                if (pressableButton != null)
+                {
+                    UpdateLabel(pressableButton.SpeechRecognitionKeyword);
+                }
+            }
         }
 
         [SerializeField]
@@ -51,16 +90,20 @@ namespace MixedReality.Toolkit.UX
             set => positionControl = value;
         }
 
-        private float canvasOffset = -10f;
-        private float nonCanvasOffset = -.004f;
+        private const float CanvasOffset = -10f;
+        private const float NonCanvasOffset = -0.004f;
+
+        protected virtual void Awake()
+        {
+            pressableButton = GetComponent<PressableButton>();
+        }
 
         /// <summary>
         /// A Unity event function that is called on the frame when a script is enabled just before any of the update methods are called the first time.
-        /// </summary> 
+        /// </summary>
         protected virtual void Start()
         {
             // Check if voice commands are enabled for this button
-            PressableButton pressableButton = gameObject.GetComponent<PressableButton>();
             if (pressableButton != null && pressableButton.AllowSelectByVoice)
             {
                 // Check if input and speech packages are present
@@ -72,23 +115,17 @@ namespace MixedReality.Toolkit.UX
                 }
 
                 SeeItSayItLabel.SetActive(true);
+                labelText = SeeItSayItLabel.GetComponentInChildren<TMP_Text>(true);
+                pressableButton.OnSpeechRecognitionKeywordChanged.AddListener(UpdateLabel);
 
-                // Children must be disabled so that they are not initially visible 
+                // Children must be disabled so that they are not initially visible
                 foreach (Transform child in SeeItSayItLabel.transform)
                 {
                     child.gameObject.SetActive(false);
                 }
 
                 // Set the label text to reflect the speech recognition keyword
-                string keyword = pressableButton.SpeechRecognitionKeyword;
-                if (keyword != null)
-                {
-                    TMP_Text labelText = SeeItSayItLabel.GetComponentInChildren<TMP_Text>(true);
-                    if (labelText != null)
-                    {
-                        labelText.text = $"Say '{keyword}'";
-                    }
-                }
+                UpdateLabel(pressableButton.SpeechRecognitionKeyword);
 
                 // If a Transform is specified, use it to reposition the object dynamically
                 if (positionControl != null)
@@ -97,26 +134,73 @@ namespace MixedReality.Toolkit.UX
                     RectTransform controlRectTransform = PositionControl.gameObject.GetComponent<RectTransform>();
 
                     // If PositionControl is a RectTransform, reposition label relative to Canvas button
-                    if (controlRectTransform != null &&  SeeItSayItLabel.transform.childCount > 0)
+                    if (controlRectTransform != null && SeeItSayItLabel.transform.childCount > 0)
                     {
                         // The parent RectTransform used to center the label
                         RectTransform canvasTransform = SeeItSayItLabel.GetComponent<RectTransform>();
 
-                        // The child RectTransform used to set the final position of the label 
+                        // The child RectTransform used to set the final position of the label
                         RectTransform labelTransform = SeeItSayItLabel.transform.GetChild(0).gameObject.GetComponent<RectTransform>();
 
                         if (labelTransform != null && canvasTransform != null)
                         {
-                            labelTransform.anchoredPosition3D = new Vector3(canvasTransform.rect.width / 2f, canvasTransform.rect.height / 2f + (controlRectTransform.rect.height /  2f * -1) + canvasOffset, canvasOffset);
+                            labelTransform.anchoredPosition3D = new Vector3(canvasTransform.rect.width / 2f, canvasTransform.rect.height / 2f + (controlRectTransform.rect.height / 2f * -1) + CanvasOffset, CanvasOffset);
                         }
                     }
                     else
                     {
-                        SeeItSayItLabel.transform.localPosition = new Vector3(PositionControl.localPosition.x, (PositionControl.lossyScale.y / 2f * -1) + nonCanvasOffset, PositionControl.localPosition.z + nonCanvasOffset);
+                        SeeItSayItLabel.transform.localPosition = new Vector3(PositionControl.localPosition.x, (PositionControl.lossyScale.y / 2f * -1) + NonCanvasOffset, PositionControl.localPosition.z + NonCanvasOffset);
                     }
                 }
+
+#if UNITY_LOCALIZATION_PRESENT
+                if (!localizedPattern.IsEmpty)
+                {
+                    localizedPattern.StringChanged += OnLocalizedPatternChanged;
+                }
+#endif
 #endif
             }
         }
+
+        protected virtual void OnDestroy()
+        {
+#if MRTK_INPUT_PRESENT && MRTK_SPEECH_PRESENT
+            if (pressableButton != null)
+            {
+                pressableButton.OnSpeechRecognitionKeywordChanged.RemoveListener(UpdateLabel);
+#if UNITY_LOCALIZATION_PRESENT
+                if (!localizedPattern.IsEmpty)
+                {
+                    localizedPattern.StringChanged -= OnLocalizedPatternChanged;
+                }
+#endif
+            }
+#endif
+        }
+
+        protected virtual void UpdateLabel(string keyword)
+        {
+#if MRTK_INPUT_PRESENT && MRTK_SPEECH_PRESENT
+            if (!string.IsNullOrWhiteSpace(keyword) && labelText != null)
+            {
+#if UNITY_LOCALIZATION_PRESENT
+                if (!localizedPattern.IsEmpty)
+                {
+                    labelText.text = localizedPattern.GetLocalizedString(keyword);
+                    return;
+                }
+#endif
+                labelText.text = string.Format(pattern, keyword);
+            }
+#endif
+        }
+
+#if MRTK_INPUT_PRESENT && MRTK_SPEECH_PRESENT && UNITY_LOCALIZATION_PRESENT
+        protected virtual void OnLocalizedPatternChanged(string value)
+        {
+            UpdateLabel(pressableButton.SpeechRecognitionKeyword);
+        }
+#endif
     }
 }