+++ org.mixedrealitytoolkit.uxcore/Tests/Runtime/SeeItSayItLabelEnablerTests.cs
@@ -4,11 +4,11 @@
 // Disable "missing XML comment" warning for tests. While nice to have, this documentation is not required.
 #pragma warning disable CS1591
 
-using System.Collections;
-using System.Text.RegularExpressions;
 using MixedReality.Toolkit.Input;
 using MixedReality.Toolkit.Input.Tests;
 using NUnit.Framework;
+using System.Collections;
+using System.Text.RegularExpressions;
 using TMPro;
 using UnityEngine;
 using UnityEngine.TestTools;
@@ -23,6 +23,9 @@ namespace MixedReality.Toolkit.UX.Runtime.Tests
         [UnityTest]
         public IEnumerator TestEnableAndSetLabel()
         {
+            GameObject testButton = SetUpButton(true, Control.None);
+            Transform label = testButton.transform.GetChild(0);
+
 #if MRTK_INPUT_PRESENT && MRTK_SPEECH_PRESENT
             if (Application.isBatchMode)
             {
@@ -30,17 +33,14 @@ namespace MixedReality.Toolkit.UX.Runtime.Tests
             }
             else
             {
-                SpeechInteractor interactor = FindObjectUtility.FindAnyObjectByType<SpeechInteractor>(true);
+                SpeechInteractor interactor = Object.FindAnyObjectByType<SpeechInteractor>(FindObjectsInactive.Include);
                 interactor.gameObject.SetActive(true);
-
-                GameObject testButton = SetUpButton(true, Control.None);
                 yield return null;
+
                 if (Application.isBatchMode)
                 {
                     LogAssert.Expect(LogType.Exception, new Regex("Speech recognition is not supported on this machine"));
                 }
-
-                Transform label = testButton.transform.GetChild(0);
 
                 Transform sublabel = label.transform.GetChild(0);
                 Assert.IsTrue(label.gameObject.activeSelf, "Label is enabled");
@@ -55,6 +55,50 @@ namespace MixedReality.Toolkit.UX.Runtime.Tests
 
             // Wait for a frame to give Unity a change to actually destroy the object
             yield return null;
+        }
+
+        [UnityTest]
+        public IEnumerator TestAutoUpdateLabel()
+        {
+            GameObject testButton = SetUpButton(true, Control.None);
+            Transform label = testButton.transform.GetChild(0);
+
+#if MRTK_INPUT_PRESENT && MRTK_SPEECH_PRESENT
+            if (Application.isBatchMode)
+            {
+                Debug.Log("Did not run SeeItSayItLabelEnablerTests, since speech is not available in batch mode.");
+            }
+            else
+            {
+                SpeechInteractor interactor = Object.FindAnyObjectByType<SpeechInteractor>(FindObjectsInactive.Include);
+                interactor.gameObject.SetActive(true);
+                yield return null;
+
+                if (Application.isBatchMode)
+                {
+                    LogAssert.Expect(LogType.Exception, new Regex("Speech recognition is not supported on this machine"));
+                }
+
+                Transform sublabel = label.transform.GetChild(0);
+                TMP_Text text = label.gameObject.GetComponentInChildren<TMP_Text>(true);
+                Assert.AreEqual(text.text, "Say 'test'", "Label text was set to voice command keyword.");
+
+                testButton.GetComponent<PressableButton>().SpeechRecognitionKeyword = "hello world";
+
+                Assert.AreEqual(text.text, "Say 'hello world'", "Label text was updated according to voice command keyword.");
+            }
+#else
+            Assert.IsTrue(!label.gameObject.activeSelf, "Did not enable label because voice commands unavailable.");
+#endif
+
+            Object.Destroy(testButton);
+            // Wait for a frame to give Unity a change to actually destroy the object
+            yield return null;
+            // The speech recognition keyword change will trigger this exception at next update when speech recognition is not supported
+            if (Application.isBatchMode)
+            {
+                LogAssert.Expect(LogType.Exception, new Regex("Speech recognition is not supported on this machine"));
+            }
         }
 
         [UnityTest]
@@ -74,6 +118,9 @@ namespace MixedReality.Toolkit.UX.Runtime.Tests
         [UnityTest]
         public IEnumerator TestPositionCanvasLabel()
         {
+            GameObject testButton = SetUpButton(true, Control.Canvas);
+            Transform label = testButton.transform.GetChild(0);
+
 #if MRTK_INPUT_PRESENT && MRTK_SPEECH_PRESENT
             if (Application.isBatchMode)
             {
@@ -81,17 +128,15 @@ namespace MixedReality.Toolkit.UX.Runtime.Tests
             }
             else
             {
-                SpeechInteractor interactor = FindObjectUtility.FindAnyObjectByType<SpeechInteractor>(true);
+                SpeechInteractor interactor = Object.FindAnyObjectByType<SpeechInteractor>(FindObjectsInactive.Include);
                 interactor.gameObject.SetActive(true);
-
-                GameObject testButton = SetUpButton(true, Control.Canvas);
                 yield return null;
+
                 if (Application.isBatchMode)
                 {
                     LogAssert.Expect(LogType.Exception, new Regex("Speech recognition is not supported on this machine"));
                 }
 
-                Transform label = testButton.transform.GetChild(0);
                 RectTransform sublabel = label.transform.GetChild(0) as RectTransform;
                 Assert.AreEqual(sublabel.anchoredPosition3D, new Vector3(10, -30, -10), "Label is positioned correctly");
                 Object.Destroy(testButton);
@@ -107,6 +152,9 @@ namespace MixedReality.Toolkit.UX.Runtime.Tests
         [UnityTest]
         public IEnumerator TestPositionNonCanvasLabel()
         {
+            GameObject testButton = SetUpButton(true, Control.NonCanvas);
+            Transform label = testButton.transform.GetChild(0);
+
 #if MRTK_INPUT_PRESENT && MRTK_SPEECH_PRESENT
             if (Application.isBatchMode)
             {
@@ -114,17 +162,15 @@ namespace MixedReality.Toolkit.UX.Runtime.Tests
             }
             else
             {
-                SpeechInteractor interactor = FindObjectUtility.FindAnyObjectByType<SpeechInteractor>(true);
+                SpeechInteractor interactor = Object.FindAnyObjectByType<SpeechInteractor>(FindObjectsInactive.Include);
                 interactor.gameObject.SetActive(true);
-
-                GameObject testButton = SetUpButton(true, Control.NonCanvas);
                 yield return null;
+
                 if (Application.isBatchMode)
                 {
                     LogAssert.Expect(LogType.Exception, new Regex("Speech recognition is not supported on this machine"));
                 }
 
-                Transform label = testButton.transform.GetChild(0);
                 Assert.AreEqual(label.transform.localPosition, new Vector3(10f, -.504f, -.004f), "Label is positioned correctly");
                 Object.Destroy(testButton);
             }
@@ -144,7 +190,7 @@ namespace MixedReality.Toolkit.UX.Runtime.Tests
             pressablebutton.AllowSelectByVoice = allowSelectByVoice;
             pressablebutton.SpeechRecognitionKeyword = "test";
 
-            // Create a label GameObject to generate 
+            // Create a label GameObject to generate
             GameObject label = new GameObject("Label");
             label.transform.SetParent(testButton.transform, false);
             label.SetActive(false);